2.1

+++ pkg/adapter/attacks.go
@@ -7,8 +7,8 @@ type AttackConfig struct {
 	Description string                 `json:"description,omitempty" yaml:"description,omitempty"`
 	Enabled     bool                   `json:"enabled,omitempty" yaml:"enabled,omitempty"`
 	TargetID    string                 `json:"target_id,omitempty" yaml:"target_id,omitempty"`
-	StartTime   int                    `json:"start_time,omitempty" yaml:"start_time,omitempty"`
-	Interval    int                    `json:"interval,omitempty" yaml:"interval,omitempty"`
+	StartTime   float64                `json:"start_time,omitempty" yaml:"start_time,omitempty"`
+	Interval    float64                `json:"interval,omitempty" yaml:"interval,omitempty"`
 	Count       int                    `json:"count,omitempty" yaml:"count,omitempty"`
 	Parameters  map[string]interface{} `json:"parameters,omitempty" yaml:"parameters,omitempty"`
 }
